consolidate avro demo
